imageAPI_docker/requirements.txt
fastapi
pillow
numpy
tensorflow-cpu
uvicorn
segmentation-models
python-multipart